audio-scribe-project-python/requirements.txt
av==12.2.0
azure-storage-blob==12.20.0
fastapi==0.111.0
moviepy==1.0.3
numpy==1.26
openai-whisper==20231117
opencv-python==4.8.0.76
pyannote.audio==3.3.1
python-dotenv==1.0.1
python-multipart==0.0.9
torch==2.2.2
transformers==4.41.2
uvicorn==0.30.1
pymongo
mimetype